digest_requirements.txt
apscheduler
exchangelib
jira
PyMySQL
peewee
python_telegram_bot
